requirements.txt
mediapipe>=0.10.8
python-osc>=1.8.3
opencv-python>=4.8.0
numpy>=1.24.0
pyfreenect2>=0.2.3 